src/main/java/com/example/demo/service/order/OrderServiceImpl.java
package com.example.demo.service.order;

import com.example.demo.service.member.entity.Member;
import com.example.demo.service.order.entity.Order;
import com.example.demo.service.member.repository.MemberRepository;
import com.example.demo.service.discount.DiscountPolicy;

import java.util.Optional;

public class OrderServiceImpl implements OrderService {

    private final MemberRepository memberRepository;
    private final DiscountPolicy discountPolicy;

    public OrderServiceImpl(MemberRepository memberRepository, DiscountPolicy discountPolicy) {
        this.memberRepository = memberRepository;
        this.discountPolicy = discountPolicy;
    }

    @Override
    public Order createOrder(Long memberId, String itemName, int itemPrice) {
        Member member = memberRepository.findById(memberId)
                .orElseThrow(() -> new IllegalArgumentException("Member not found with ID: " + memberId));
        int discountPrice = discountPolicy.discount(member, itemPrice);

        return new Order(memberId, itemName, itemPrice, discountPrice);
    }
}

